add filter category select
Categories colud not be filtered by user groups. Could confuse the guide editors upon selecting categories for a new or existing guide.

Now the selection of guide groups is filtered by the user group that a user is in. Categories that are not tagged to a group will always display amongst the selection option.